wip on cleaning things up

+++ src/plugins/plugin-manager.ts
@@ -1,7 +1,12 @@
-import { existsSync } from 'std/fs/exists.ts';
-import * as path from 'std/path/mod.ts';
-import { ArchitectPlugin, PluginArchitecture, PluginBundleType, PluginPlatform } from './plugin-types.ts';
-import PluginUtils from './plugin-utils.ts';
+import { existsSync } from "std/fs/exists.ts";
+import * as path from "std/path/mod.ts";
+import {
+  ArchitectPlugin,
+  PluginArchitecture,
+  PluginBundleType,
+  PluginPlatform,
+} from "./plugin-types.ts";
+import PluginUtils from "./plugin-utils.ts";
 
 type Dictionary<T> = { [key: string]: T };
 
@@ -27,7 +32,10 @@ export default class PluginManager {
     return this.ARCHITECTUREMAP[Deno.build.arch];
   }
 
-  private static async removeOldPluginVersions(pluginDirectory: string, plugin: ArchitectPlugin) {
+  private static async removeOldPluginVersions(
+    pluginDirectory: string,
+    plugin: ArchitectPlugin,
+  ) {
     if (!existsSync(pluginDirectory)) {
       return;
     }
@@ -57,33 +65,58 @@ export default class PluginManager {
       throw new Error(`Unable to find version ${version} of ${ctor.name}`);
     }
     const pluginDirectory = configDirectory;
-    const currentPluginDirectory = path.join(pluginDirectory, `/${plugin.name}`);
+    const currentPluginDirectory = path.join(
+      pluginDirectory,
+      `/${plugin.name}`,
+    );
     const versionPath = path.join(currentPluginDirectory, `/${version}`);
 
     await this.removeOldPluginVersions(currentPluginDirectory, plugin);
     await Deno.mkdir(versionPath, { recursive: true });
 
-    const binary = PluginUtils.getBinary(plugin.versions[version], this.getPlatform(), this.getArchitecture());
+    const binary = PluginUtils.getBinary(
+      plugin.versions[version],
+      this.getPlatform(),
+      this.getArchitecture(),
+    );
     const downloadedFilePath = path.join(
       versionPath,
-      `/${plugin.name}.${binary.bundleType === PluginBundleType.ZIP ? 'zip' : 'tar.gz'}`,
+      `/${plugin.name}.${
+        binary.bundleType === PluginBundleType.ZIP ? "zip" : "tar.gz"
+      }`,
     );
 
     const executablePath = path.join(versionPath, `/${binary.executablePath}`);
-    // if (!existsSync(executablePath)) { // TODO: comment about why this was changed - related to https://github.com/denoland/deno_std/issues/1216, https://github.com/denoland/deno_std/issues/2494
-      await PluginUtils.downloadFile(binary.url, downloadedFilePath, binary.sha256);
-      await PluginUtils.extractFile(downloadedFilePath, versionPath, binary.bundleType);
-      try {
-        await Deno.chmod(executablePath, 0o755);
-      } catch {} // TODO: add a real error or something here
-      try {
-        await Deno.remove(downloadedFilePath);
-      } catch {} // TODO: add a real error or something here
-    // }
+    let directory_exists = false;
+    try {
+      if (existsSync(executablePath)) {
+        directory_exists = true;
+      }
+    } catch {
+      // ignore error if directory doesn't exist as existsSync will throw an error - https://github.com/denoland/deno_std/issues/1216, https://github.com/denoland/deno_std/issues/2494
+    }
+    if (!directory_exists) {
+      await PluginUtils.downloadFile(
+        binary.url,
+        downloadedFilePath,
+        binary.sha256,
+      );
+      await PluginUtils.extractFile(
+        downloadedFilePath,
+        versionPath,
+        binary.bundleType,
+      );
+      await Deno.chmod(executablePath, 0o755);
+      await Deno.remove(downloadedFilePath);
+    }
 
     await plugin.setup(
       versionPath,
-      PluginUtils.getBinary(plugin.versions[version], this.getPlatform(), this.getArchitecture()),
+      PluginUtils.getBinary(
+        plugin.versions[version],
+        this.getPlatform(),
+        this.getArchitecture(),
+      ),
     );
 
     this.plugins[id] = plugin;
